blogilistan laajennus step9

+++ blogilista/controllers/blogs.js
@@ -55,6 +55,23 @@ blogsRouter.post('/', async (request, response) => {
 
 
 blogsRouter.delete('/:id', async (request, response) => {
+    if (!request.token) {
+        return response.status(401).json({error: 'token missing or invalid'})
+}
+let decodedToken
+try {
+    decodedToken = jwt.verify(request.token, process.env.SECRET)
+} catch (error) {
+    return response.status(401).json({error: 'token missing or invalid'})
+}
+const blog = await Blog.findById(request.params.id)
+if(!blog) {
+    return response.status(404).json({error: 'blog not found'})
+}
+if (!blog.user || blog.user.toString() !== decodedToken.id.toString()) {
+    return response.status(403).json({error: 'only the user who added the blog can delete it'})
+}
+
     await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()})
